fix leak

+++ bin/test-runner.sh
@@ -1,24 +1,34 @@
 #!/bin/bash
 
 # Test runner for running jest one by one and collect the overall coverage at the end
-#
-# To run a single test (or a pattern) you can use:
-# $ ./node_modules/.bin/jest --config ./packages/THE_PACKAGE/jest.config.js ./packages/THE_PACKAGE/path/tp/file/or/dir
 
 jest=./node_modules/.bin/jest
 nyc=./node_modules/.bin/nyc
 
-rm -rf ./coverage
+# Run on the given path
+if [ -f "$1" ] || [ -d "$1" ] ; then
+  path=${1}
+  pkg=$(echo "${path}" | sed -n -e 's/^.*packages//p' | awk -F "/" '{print $2}')
+  shift
 
+  ${jest} "${path}" --config "./packages/${pkg}/jest.config.js" --color "$@"
+  exit $?
+fi
+
+# Run on all
+rm -rf ./coverage
 for dir in ./packages/* ; do
   pkg=$(basename "${dir}")
 
-  if [ ]
-  #${jest} ./packages/"${pkg}" --config "./packages/${pkg}/jest.config.js" --color "$@"
-  # Comment the previous line and uncomment this line to check for memory leak
-   node --expose-gc --trace-warnings ${jest} ./packages/"${pkg}" --config "./packages/${pkg}/jest.config.js" -i --runInBand --logHeapUsage
-
-  #mv ./coverage/coverage-final.json "./coverage/${pkg}.json"
+  if [ -z "$EXPOSE_GC" ] ; then
+    ${jest} ./packages/"${pkg}" --config "./packages/${pkg}/jest.config.js" --color --coverage "$@"
+    mv ./coverage/coverage-final.json "./coverage/${pkg}.json"
+  else
+    node --expose-gc --trace-warnings ${jest} ./packages/"${pkg}" --config "./packages/${pkg}/jest.config.js" -i --runInBand --logHeapUsage
+  fi
 done
 
-#${nyc} report -t ./coverage --report-dir ./coverage --reporter=html --reporter=lcov
+# Combine reports
+if [ "$(ls -A ./coverage)" ] ; then
+  ${nyc} report -t ./coverage --report-dir ./coverage --reporter=html --reporter=lcov
+fi
